+++ src/main/kotlin/com/github/biomejs/intellijbiome/services/BiomeServerService.kt
@@ -1,18 +1,21 @@
 package com.github.biomejs.intellijbiome.services
 
-import com.intellij.openapi.components.Service
-import com.intellij.openapi.project.Project
 import com.github.biomejs.intellijbiome.BiomeBundle
 import com.github.biomejs.intellijbiome.lsp.BiomeLspServerSupportProvider
 import com.intellij.notification.NotificationGroupManager
 import com.intellij.notification.NotificationType
+import com.intellij.openapi.components.Service
+import com.intellij.openapi.project.Project
 import com.intellij.platform.lsp.api.LspServerManager
 
 @Service(Service.Level.PROJECT)
 class BiomeServerService(private val project: Project) {
-
     fun restartBiomeServer() {
         LspServerManager.getInstance(project).stopAndRestartIfNeeded(BiomeLspServerSupportProvider::class.java)
+    }
+
+    fun stopBiomeServer() {
+        LspServerManager.getInstance(project).stopServers(BiomeLspServerSupportProvider::class.java)
     }
 
     fun notifyRestart() {