+++ src/test/kotlin/com/github/biomejs/intellijbiome/pages/StatusBar.kt
@@ -4,20 +4,28 @@ import com.intellij.remoterobot.RemoteRobot
 import com.intellij.remoterobot.data.RemoteComponent
 import com.intellij.remoterobot.fixtures.*
 import com.intellij.remoterobot.search.locators.byXpath
-import com.intellij.remoterobot.utils.waitFor
 import java.time.Duration
+
 fun RemoteRobot.statusBar(function: StatusbarFrame.() -> Unit) {
-	find<StatusbarFrame>(timeout = Duration.ofSeconds(10)).apply(function)
+    find<StatusbarFrame>(timeout = Duration.ofSeconds(10)).apply(function)
 }
 
 @FixtureName("Statusbar frame")
 @DefaultXpath("IdeStatusBarImpl type", "//div[@class='IdeStatusBarImpl']")
 class StatusbarFrame(remoteRobot: RemoteRobot, remoteComponent: RemoteComponent) :
-	CommonContainerFixture(remoteRobot, remoteComponent) {
+    CommonContainerFixture(remoteRobot, remoteComponent) {
 
-	val statusBarPanel get() = find<ContainerFixture>(byXpath("StatusBarPanel", "//div[@class='StatusBarPanel'][.//div[@class='CodeStyleStatusBarPanel']]"))
-	fun byContainsText(text: String) = byXpath("text $text", "//div[contains(@text,'$text') and @class='WithIconAndArrows']")
+    val statusBarPanel
+        get() = find<ContainerFixture>(
+            byXpath(
+                "StatusBarPanel",
+                "//div[@class='StatusBarPanel'][.//div[@class='CodeStyleStatusBarPanel']]"
+            )
+        )
 
-	val text: String
-		get() = callJs("component.getText();")
+    fun byContainsText(text: String): ComponentFixture =
+        find<ComponentFixture>(byXpath("//div[contains(@text,'$text') and @text.key='biome.widget.version']"))
+
+    val text: String
+        get() = callJs("component.getText();")
 }