+++ CHANGELOG.md
@@ -2,3 +2,26 @@
 
 All notable changes to this project will be documented in this file.
 
+## [unreleased]
+
+### Bug Fixes
+
+- Added missing double quotes ([#377](https://github.com/biomejs/biome-intellij/pull/377))
+- Use node interpreter to run commands ([#416](https://github.com/biomejs/biome-intellij/pull/416))
+- Binary resolution on windows ([#556](https://github.com/biomejs/biome-intellij/pull/556))
+- Binary resolution execution sequence ([#601](https://github.com/biomejs/biome-intellij/pull/601))
+- Remove build range ([#1093](https://github.com/biomejs/biome-intellij/pull/1093))
+- Auto-save race condition ([#26](https://github.com/biomejs/biome-intellij/pull/26))
+
+### Documentation
+
+- Add contribution guide ([#2](https://github.com/biomejs/biome-intellij/pull/2))
+
+### Features
+
+- IntelliJ Platform LSP ([#185](https://github.com/biomejs/biome-intellij/pull/185))
+- Manual config path specifying ([#660](https://github.com/biomejs/biome-intellij/pull/660))
+- Add onSave actions
+- Improved command line execution mode (node & binary)
+- Use biome check ([#28](https://github.com/biomejs/biome-intellij/pull/28))
+